Move vfs to ext/dynamic mod to prevent issues with wasm
