+++ core/ext/dynamic.rs
@@ -1,6 +1,6 @@
 use crate::{Connection, LimboError};
 use libloading::{Library, Symbol};
-use limbo_ext::{ExtensionApi, ExtensionApiRef, ExtensionEntryPoint};
+use limbo_ext::{ExtensionApi, ExtensionApiRef, ExtensionEntryPoint, VfsImpl};
 use std::{
     rc::Rc,
     sync::{Arc, Mutex, OnceLock},
@@ -8,11 +8,23 @@ use std::{
 
 type ExtensionStore = Vec<(Arc<Library>, ExtensionApiRef)>;
 static EXTENSIONS: OnceLock<Arc<Mutex<ExtensionStore>>> = OnceLock::new();
+
 pub fn get_extension_libraries() -> Arc<Mutex<ExtensionStore>> {
     EXTENSIONS
         .get_or_init(|| Arc::new(Mutex::new(Vec::new())))
         .clone()
 }
+
+#[derive(Clone, Debug)]
+pub struct VfsMod {
+    pub ctx: *const VfsImpl,
+}
+
+unsafe impl Send for VfsMod {}
+unsafe impl Sync for VfsMod {}
+type Vfs = (String, Arc<VfsMod>);
+
+static VFS_MODULES: OnceLock<Mutex<Vec<Vfs>>> = OnceLock::new();
 
 impl Connection {
     pub fn load_extension<P: AsRef<std::ffi::OsStr>>(
@@ -43,4 +55,32 @@ impl Connection {
             ))
         }
     }
+}
+
+pub(crate) fn add_vfs_module(name: String, vfs: Arc<VfsMod>) {
+    let mut modules = VFS_MODULES
+        .get_or_init(|| Mutex::new(Vec::new()))
+        .lock()
+        .unwrap();
+    if !modules.iter().any(|v| v.0 == name) {
+        modules.push((name, vfs));
+    }
+}
+
+pub fn list_vfs_modules() -> Vec<String> {
+    VFS_MODULES
+        .get_or_init(|| Mutex::new(Vec::new()))
+        .lock()
+        .unwrap()
+        .iter()
+        .map(|v| v.0.clone())
+        .collect()
+}
+
+pub(crate) fn get_vfs_modules() -> Vec<Vfs> {
+    VFS_MODULES
+        .get_or_init(|| Mutex::new(Vec::new()))
+        .lock()
+        .unwrap()
+        .clone()
 }
